Fix warnings in Visual Studio examples due to variables with wrong type

+++ examples/VisualStudio/Example_Floating/Example_Floating.cpp
@@ -47,7 +47,7 @@ void activate(char const* machine_code)
 			std::cout << "Cryptolens error: " << e.get_reason() << std::endl;
 		}
 		else if (e.get_subsystem() == Subsystem::RequestHandler) {
-			int code = e.get_extra();
+			size_t code = e.get_extra();
 			std::cout << "Error connecting to the server: error code: " << code << std::endl;
 		}
 		else {
